backend/requirements.txt
fastapi
uvicorn[standard]
torch --extra-index-url https://download.pytorch.org/whl/cu121
diffusers
transformers
accelerate
safetensors
Pillow
imageio[ffmpeg]
opencv-python